switching everything to utils folder and clean coding
