+++ utils/helpers-sever.ts
@@ -1,0 +1,63 @@
+// FIREBASE ADMIN
+import { firebaseAuth, firestore } from "./firebase-admin";
+// STRIPE
+import { stripe } from "./stripe";
+import Stripe from "stripe";
+
+/* ----------USER & AUTHENTICATION --------------*/
+export async function validateUser(req: Request) {
+  // grab auth token from http headers
+  const headers = new Headers(req.headers);
+  const bearerToken = headers.get('authorization')?.split('Bearer ')[1] || '';
+
+  // look up the user on firebase
+  const decodedToken = await firebaseAuth.verifyIdToken(bearerToken);
+  const user = { uid: decodedToken.uid, email: decodedToken.email };
+
+  // if nothing shows up, user doesn't exist and we should throw an error
+  if (!user.uid) {
+    throw new Error('You must be logged in to make this request');
+  }
+
+  return user;
+}
+
+export async function getOrCreateCustomer(uid: string, params?: Stripe.CustomerCreateParams) {
+  // extract user info
+  const userDocument = await firestore.collection('users').doc(uid).get();
+  const { stripeCustomerId, email } = userDocument.data() ?? {};
+
+  if (!stripeCustomerId) {
+    // create new stripe customer if their details don't exist on firestore
+    const customer = await stripe.customers.create({
+      email: email,
+      metadata: {
+        firebaseUID: uid
+      },
+      ...params
+    })
+
+    await userDocument.ref.update({stripeCustomerId: customer.id });
+    return customer;
+  } else {
+    // otherwise, retrieve customer with firebaseUID
+    return await stripe.customers.retrieve(stripeCustomerId);
+  }
+}
+
+/* --------------- Firestore --------------*/
+export async function updateProductOnFirestore(product: Stripe.Product) {
+  const { id, active, name, description, metadata } = product;
+  const response = await firestore.collection('products').doc(id).set({
+    id,
+    active,
+    name,
+    description,
+    image: product.images?.[0] ?? null,
+    metadata
+  }, { merge: true });
+
+  if (!response) throw new Error('Failed to create product on firestore');
+
+  console.log(`Inserted/updated product [${id}]`)
+}
